requirement.txt
fastapi
uvicorn
langchain
langchain-community
python-dotenv
requests
groq